updates reportes backend

+++ ERP/Pages/Cotizacion/CotizacionArticulo/CotizacionArticuloList.cshtml.cs
@@ -35,9 +35,9 @@ namespace ERP.Pages.Cotizacion.CotizacionArticulo
                             cantidad_inventario = reader.GetInt32(2).ToString(),
                             estado_articulo = reader.GetString(3), // Convertimos a "Activo"/"Inactivo"
                             descripcion_articulo = reader.GetString(4),
-                            peso_articulo = reader.GetDecimal(5).ToString("F2"),
-                            costo_articulo = reader.GetDecimal(6).ToString("F2"),
-                            precio_articulo = reader.GetDecimal(7).ToString("F2"),
+                            peso_articulo = reader.GetDouble(5).ToString("F2"),
+                            costo_articulo = reader.GetDouble(6).ToString("F2"),
+                            precio_articulo = reader.GetDouble(7).ToString("F2"),
                             numero_cotizacion = reader.GetInt32(8).ToString(),
                             cantidad_cotizada = reader.GetInt32(9).ToString(),
                             monto_cotizacion = reader.GetDouble(10).ToString("F2")
